# gen/docs/moving.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "moving.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "MovingService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/orders/create": {
      "post": {
        "operationId": "MovingService_CreateOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateOrderRequest"
            }
          }
        ],
        "tags": [
          "MovingService"
        ]
      }
    },
    "/v1/orders/get": {
      "get": {
        "operationId": "MovingService_AllOrders",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AllOrdersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "MovingService"
        ]
      }
    },
    "/v1/orders/update": {
      "post": {
        "operationId": "MovingService_UpdateOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdateOrderRequest"
            }
          }
        ],
        "tags": [
          "MovingService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AllOrdersResponse": {
      "type": "object",
      "properties": {
        "Orders": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Order"
          }
        }
      }
    },
    "v1CreateOrderRequest": {
      "type": "object",
      "properties": {
        "PropertySize": {
          "$ref": "#/definitions/v1PropertySize"
        },
        "OrderStatus": {
          "$ref": "#/definitions/v1OrderStatus"
        },
        "MoveDate": {
          "type": "string",
          "format": "date-time"
        },
        "Name": {
          "type": "string"
        },
        "Email": {
          "type": "string"
        },
        "Phone": {
          "type": "string"
        },
        "MoveFrom": {
          "type": "string"
        },
        "MoveTo": {
          "type": "string"
        },
        "AdditionalInfo": {
          "type": "string"
        }
      }
    },
    "v1CreateOrderResponse": {
      "type": "object",
      "properties": {
        "Order": {
          "$ref": "#/definitions/v1Order"
        }
      }
    },
    "v1Order": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uint64"
        },
        "PropertySize": {
          "$ref": "#/definitions/v1PropertySize"
        },
        "OrderStatus": {
          "$ref": "#/definitions/v1OrderStatus"
        },
        "MoveDate": {
          "type": "string",
          "format": "date-time"
        },
        "Name": {
          "type": "string"
        },
        "Email": {
          "type": "string"
        },
        "Phone": {
          "type": "string"
        },
        "MoveFrom": {
          "type": "string"
        },
        "MoveTo": {
          "type": "string"
        },
        "AdditionalInfo": {
          "type": "string"
        }
      }
    },
    "v1OrderStatus": {
      "type": "string",
      "enum": [
        "ORDER_STATUS_UNKNOWN",
        "ORDER_STATUS_CREATED",
        "ORDER_STATUS_REJECTED",
        "ORDER_STATUS_IN_PROGRESS",
        "ORDER_STATUS_DONE"
      ],
      "default": "ORDER_STATUS_UNKNOWN"
    },
    "v1PropertySize": {
      "type": "string",
      "enum": [
        "PROPERTY_SIZE_UNKNOWN",
        "PROPERTY_SIZE_STUDIO",
        "PROPERTY_SIZE_1_BEDROOM",
        "PROPERTY_SIZE_2_BEDROOMS",
        "PROPERTY_SIZE_3_BEDROOMS",
        "PROPERTY_SIZE_4_PLUS_BEDROOMS",
        "PROPERTY_SIZE_COMMERCIAL"
      ],
      "default": "PROPERTY_SIZE_UNKNOWN"
    },
    "v1UpdateOrderRequest": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uint64"
        },
        "PropertySize": {
          "$ref": "#/definitions/v1PropertySize"
        },
        "OrderStatus": {
          "$ref": "#/definitions/v1OrderStatus"
        },
        "MoveDate": {
          "type": "string",
          "format": "date-time"
        },
        "Name": {
          "type": "string"
        },
        "Email": {
          "type": "string"
        },
        "Phone": {
          "type": "string"
        },
        "MoveFrom": {
          "type": "string"
        },
        "MoveTo": {
          "type": "string"
        },
        "AdditionalInfo": {
          "type": "string"
        }
      }
    }
  }
}